requirements.txt
# Web framework
Flask==2.3.3

# Web scraping
selenium==4.14.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1

# Natural language processing
textblob==0.17.1
nltk==3.8.1

# Data processing
python-dateutil==2.8.2

# Utilities
tqdm==4.66.1